WIP: nice plugin interface
TODO: destroy the gl.GenTextures
